last

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: kel-malt <kel-malt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/04 13:58:11 by kel-malt          #+#    #+#             */
-/*   Updated: 2024/03/24 18:20:33 by kel-malt         ###   ########.fr       */
+/*   Updated: 2024/03/25 12:01:08 by kel-malt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,9 +27,9 @@ void	putarg(const char arg_type, va_list args, int *len)
 	else if (arg_type == 'u')
 		ft_put_unsignedint(va_arg(args, unsigned int), len);
 	else if (arg_type == 'x')
-		ft_put_hex_lower(va_arg(args, unsigned long), len);
+		ft_put_hex_lower(va_arg(args, unsigned int), len);
 	else if (arg_type == 'X')
-		ft_put_hex_upper(va_arg(args, unsigned long), len);
+		ft_put_hex_upper(va_arg(args, unsigned int), len);
 	else if (arg_type == '%')
 		ft_put_char('%', len);
 }
